Cookiecutter/src/models/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_pytorch_select=0.1=cpu_0
appnope=0.1.2=py39hecd8cb5_1001
argon2-cffi=20.1.0=py39h9ed2024_1
arrow=1.1.0=pyhd8ed1ab_1
async_generator=1.10=pyhd3eb1b0_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.9.3=pyha847dfd_0
binaryornot=0.4.4=py_1
blas=1.0=mkl
bleach=3.3.0=pyhd3eb1b0_0
brotlipy=0.7.0=py39h9ed2024_1003
bzip2=1.0.8=h1de35cc_0
ca-certificates=2021.5.25=hecd8cb5_1
certifi=2021.5.30=py39hecd8cb5_0
cffi=1.14.5=py39h2125817_0
chardet=4.0.0=py39hecd8cb5_1003
click=8.0.1=py39h6e9494a_0
conda=4.10.1=py39hecd8cb5_1
conda-build=3.21.4=py39hecd8cb5_0
conda-package-handling=1.7.3=py39h9ed2024_1
cookiecutter=1.7.3=pyh6c4a22f_0
cryptography=3.4.7=py39h2fd3fbb_0
cycler=0.10.0=py39hecd8cb5_0
decorator=5.0.9=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=py39hecd8cb5_0
filelock=3.0.12=pyhd3eb1b0_1
freetype=2.10.4=ha233b18_0
glob2=0.7=pyhd3eb1b0_0
icu=58.2=h0a44026_3
idna=2.10=pyhd3eb1b0_0
importlib-metadata=3.10.0=py39hecd8cb5_0
importlib_metadata=3.10.0=hd3eb1b0_0
intel-openmp=2019.4=233
ipykernel=5.3.4=py39h01d92e1_0
ipython=7.22.0=py39h01d92e1_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pyhd3eb1b0_1
jedi=0.17.2=py39hecd8cb5_1
jinja2=2.11.3=pyh44b312d_0
jinja2-time=0.2.0=py_2
jpeg=9b=he5867d9_2
jsonschema=3.2.0=py_2
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py39hecd8cb5_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.3.1=py39h23ab428_0
lcms2=2.12=hf1fd2bf_0
libarchive=3.4.2=haa3ed63_0
libcxx=10.0.0=1
libffi=3.3=hb1e8313_2
libiconv=1.16=h1de35cc_0
liblief=0.10.1=h23ab428_1
libmklml=2019.0.5=0
libpng=1.6.37=ha441bb4_0
libsodium=1.0.18=h1de35cc_0
libtiff=4.2.0=h87d7836_0
libwebp-base=1.2.0=h9ed2024_0
libxml2=2.9.10=h7cdb67c_3
llvm-openmp=10.0.0=h28b9765_0
lz4-c=1.9.3=h23ab428_0
markupsafe=1.1.1=py39hcbf5805_3
matplotlib=3.3.4=py39hecd8cb5_0
matplotlib-base=3.3.4=py39h8b3ea08_0
mistune=0.8.4=py39h9ed2024_1000
mkl=2019.4=233
mkl-service=2.3.0=py39h9ed2024_0
mkl_fft=1.3.0=py39ha059aab_0
mkl_random=1.0.2=py39h16bde0e_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.0.7=py39hecd8cb5_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.2=h0a44026_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
ninja=1.10.2=hf7b0b51_1
notebook=6.4.0=py39hecd8cb5_0
numpy=1.19.2=py39he57783f_0
numpy-base=1.19.2=py39hde55871_0
olefile=0.46=py_0
openssl=1.1.1k=h9ed2024_0
packaging=20.9=pyhd3eb1b0_0
pandoc=2.12=hecd8cb5_0
pandocfilters=1.4.3=py39hecd8cb5_1
parso=0.7.0=py_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.2.0=py39h5270095_0
pkginfo=1.7.0=py39hecd8cb5_0
poyo=0.5.0=py_0
prometheus_client=0.10.1=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyh06a4308_0
psutil=5.8.0=py39h9ed2024_1
ptyprocess=0.7.0=pyhd3eb1b0_2
py-lief=0.10.1=py39h23ab428_1
pycosat=0.6.3=py39h9ed2024_0
pycparser=2.20=py_2
pygments=2.9.0=pyhd3eb1b0_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyrsistent=0.17.3=py39h9ed2024_0
pysocks=1.7.1=py39hecd8cb5_0
python=3.9.1=h88f2d9e_2
python-dateutil=2.8.1=pyhd3eb1b0_0
python-libarchive-c=2.9=pyhd3eb1b0_1
python-slugify=5.0.2=pyhd8ed1ab_0
python.app=3=py39h9ed2024_0
python_abi=3.9=1_cp39
pytorch=1.7.1=cpu_py39h7e2095a_0
pytz=2021.1=pyhd3eb1b0_0
pyyaml=5.4.1=py39h9ed2024_1
pyzmq=20.0.0=py39h23ab428_1
readline=8.1=h9ed2024_0
requests=2.25.1=pyhd3eb1b0_0
ripgrep=12.1.1=0
ruamel_yaml=0.15.100=py39h9ed2024_0
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=52.0.0=py39hecd8cb5_0
six=1.15.0=py39hecd8cb5_0
soupsieve=2.2.1=pyhd3eb1b0_0
sqlite=3.35.4=hce871da_0
terminado=0.9.4=py39hecd8cb5_0
testpath=0.4.4=pyhd3eb1b0_0
text-unidecode=1.3=py_0
tk=8.6.10=hb0a8c7a_0
torchvision=0.8.2=cpu_py39hde629fd_0
tornado=6.1=py39h9ed2024_0
tqdm=4.59.0=pyhd3eb1b0_1
traitlets=5.0.5=pyhd3eb1b0_0
typing-extensions=3.7.4.3=hd3eb1b0_0
typing_extensions=3.7.4.3=pyh06a4308_0
tzdata=2020f=h52ac0ba_0
unidecode=1.2.0=pyhd8ed1ab_0
urllib3=1.26.4=pyhd3eb1b0_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py39hecd8cb5_1
widgetsnbextension=3.5.1=py39hecd8cb5_0
xz=5.2.5=h1de35cc_0
yaml=0.2.5=haf1e3a3_0
zeromq=4.3.4=h23ab428_0
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.9=h322a384_0
